+++ src/components/datepicker/DatePickerInput.js
@@ -77,10 +77,12 @@ class DatePickerInput extends PureComponent {
       );
     }
 
-    onClick(event);
+    onClick && onClick(event);
   };
 
   handlePopoverClose = () => {
+    const { onBlur } = this.props;
+    onBlur && onBlur({ relatedTarget: null });
     this.setState({ isPopoverActive: false });
   };
 
@@ -164,6 +166,7 @@ class DatePickerInput extends PureComponent {
               overflowX="hidden"
               paddingHorizontal={3}
               paddingVertical={3}
+              tabIndex="0"
             >
               {footer}
             </Box>
@@ -191,6 +194,8 @@ DatePickerInput.propTypes = {
   locale: PropTypes.string,
   /** Callback function that is fired when the date has changed. */
   onChange: PropTypes.func,
+  /** Callback function that is fired when the popover with the calendar gets closed (unfocused) */
+  onBlur: PropTypes.func,
   /** Object with props for the Popover component. */
   popoverProps: PropTypes.object,
   /** The current selected date. */